feat(api): continue the split (balance)
- Add rpc.balance.cpp/.hpp that represents the mm2 balance rpc call
- Remove balance rpc from mm2.cpp/.hpp
- Add the sources to the CMakeLists.txt
- Fix compilation on windows with kill
